raise exception when input file is empty
This can happen when the input file is a external list, and need to be
downloaded, but the download failed, leading to an empty geo*.dat file.

If this happens, the route policy of V2Ray just fails sliently. By
raising a exception, the python script exits immediately, without
touching or overwriting output file at all.